Fix format string for timeout_t on 64-bit platforms
timeout_t is 'long int' which is 64-bit on LP64 platforms (like aarch64).
The %d format specifier expects 'int', causing -Wformat warnings that
become errors with -Werror.

Use %ld instead to match the actual argument type.

+++ src/XrdClPelican/PelicanFile.cc
@@ -315,7 +315,7 @@ File::Open(const std::string      &url,
     }
 
     auto ts = GetHeaderTimeout(timeout);
-    m_logger->Debug(kLogXrdClPelican, "Opening %s (with timeout %d)", m_url.c_str(), timeout);
+    m_logger->Debug(kLogXrdClPelican, "Opening %s (with timeout %ld)", m_url.c_str(), static_cast<long int>(timeout));
 
     std::unique_ptr<XrdCl::ResponseHandler> wrapped_handler(
         new DirectorCacheResponseHandler<XrdClCurl::OpenResponseInfo, XrdClCurl::OpenResponseInfo>(
